requirements.txt
minepy
absl-py==0.7.1
astor==0.8.0
beautifulsoup4==4.7.1
bs4==0.0.1
certifi==2019.3.9
chardet==3.0.4
cycler==0.10.0
gast==0.2.2
google-pasta==0.1.7
grpcio==1.21.1
h5py==2.9.0
idna==2.8
joblib==0.13.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lxml==4.3.3
Markdown==3.1.1
matplotlib==3.0.3
mpl-finance==0.10.0
numpy==1.16.3
opt-einsum==3.1.0
pandas==0.24.2
pandas-datareader==0.7.0
protobuf==3.8.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.21.0
scikit-learn==0.21.2
scipy==1.2.1
six==1.12.0
sklearn==0.0
soupsieve==1.9.1
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
urllib3==1.24.3
Werkzeug==0.15.4
wrapt==1.11.1
